backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==21.2.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
google-auth==2.37.0
requests==2.32.3
python-dotenv==1.0.1
